\name{getAlleleSPDF}
\alias{getAlleleSPDF}
\title{getAlleleSPDF}
\usage{
getAlleleSPDF(data)
}
\description{
Make a spatial polygons data frame for IDW analysis based on sample. This is designed for cross validation.
}
\arguments{
\item{data}{A character vector loaded from a csv file.}
}
\value{
A spatial polygons data frame object.
}
