\name{getPopulationAlleleSPDF}
\alias{getPopulationAlleleSPDF}
\title{getPopulationAlleleSPDF}
\usage{
getPopulationAlleleSPDF(data,pdata)
}
\description{
Make a spatial polygons data frame for IDW analysis based on whole population.
}
\arguments{
\item{data}{A character vector loaded from a csv file or resampled one from pdata.}
\item{pdata}{A character vector loaded from a csv file.}
}
\value{
A spatial polygons data frame object.
}
