\name{polygons2centerpoints}
\alias{polygons2centerpoints}
\title{polygons2centerpoints}
\usage{
polygons2centerpoints(spdf)
}
\description{
Convert a spatial polygons data frame into points data of the polygon centers for SPDE and IDW analysis.
}
\arguments{
\item{spdf}{A spatial polygons data frame object.}
}
\value{
A matrix of two cloumns which correspond to the coordinates.
}
