\name{resampleFreq}
\alias{resampleFreq}
\title{resampleFreq}
\usage{
resampleFreq(model,n)
}
\description{
Resample allele frequency from posterior distribution.
}
\arguments{
\item{model}{A list of calculated models.}
\item{n}{An integer of the number of size of resampling.}
}
\value{
A list of resamples.
}
