\name{alleleAtlas-internal}
\alias{meshProj2Rectangles}
\alias{makeMinWindow}
\alias{interpreteBase}
\alias{seq.dist}
\alias{calcHTAE}
\alias{calcLogLoss}
\alias{leaveOneOut}
\alias{validateFrequency}
\alias{resampleFreq}
\title{Internal alleleAtlas Functions}
\description{
 Internal alleleAtlas functions
}
\details{
These are not to be called by the user.
}
\keyword{ts}
